feat: validation for story settings
