+++ src/modules/editor/components/SlateEditor.tsx
@@ -20,10 +20,11 @@ import {
   getStoriesFile,
   saveStoriesFile,
   generateRandomId,
+  getStoryFile,
 } from '../../../utils';
 import { Story } from '../../../types';
 import { Content } from '../../publicStory/components/PublicStory';
-import { StorySettings } from '../containers/StorySettings';
+import { StorySettings } from './StorySettings';
 import { SlateEditorSideMenu } from './SlateEditorSideMenu';
 import { SlateEditorHoverMenu } from './SlateEditorHoverMenu';
 import { SlateEditorToolbar } from './SlateEditorToolbar';
@@ -435,13 +436,14 @@ export const SlateEditor = ({
     next();
   };
 
-  const handleSave = async (storyParam?: Partial<Story>) => {
+  const handleSave = async () => {
     setLoadingSave(true);
     try {
+      const storyFile = (await getStoryFile(story.id)) as Story;
       const content = value.toJSON();
       const updatedStory: Story = {
-        ...story,
-        ...storyParam,
+        ...storyFile,
+        title: story.title,
         content,
         updatedAt: Date.now(),
       };
@@ -581,8 +583,6 @@ export const SlateEditor = ({
           story={story}
           open={settingsOpen}
           onClose={handleCloseSettings}
-          onChangeStoryField={onChangeStoryField}
-          onSave={handleSave}
         />
 
         <FullScreenDialog
